Struct quasi::Span

pub struct Span {
    pub lo: BytePos,
    pub hi: BytePos,
    pub expn_id: ExpnId,
}

Unstable

Spans represent a region of code, used for error reporting. Positions in spans are absolute positions from the beginning of the codemap, not positions relative to FileMaps. Methods on the CodeMap can be used to relate spans back to the original source.

Fields

lo
Unstable

hi
Unstable

expn_id
Unstable

Information about where the macro came from, if this piece of code was created by a macro expansion.